Simplified the approach to making the app responsive by utilising MaterialUI's Hidden component.
Now, instead of rendering two different component, one for mobile
view and another for desktop view, that get hidden based on the
desktop width, the app uses the Hidden component to hide either
the TweetInput component or the ThreadViewer component based on the
value of the 'editing' state variable, which can be toggled when in
mobile view using the 'Show Thread' and 'Edit Thread' buttons.

In desktop view, both the TweetInput and the ThreadViewer are visible
and the 'Show Thread' and 'Edit Thread' buttons are hidden, so the
'editing' state cannot be changed.