Started implementing splitting the tweets at full sentences but it still needs improvements to be working properly.
Currently, the function returns once it encounters a sentence that
doesn't fit within a tweet, even if there are other sentences later
that would fit in their own tweets.

It needs to be fixed so, when that happens, it would just add that
long sentence as it is and leave it for the rest of the logic to
handle breaking it into chunks that would fit in a tweet.