Separated all calls to the backend in their own module that is imported inside the UI modules.
Implemented a check when the user clicks the 'Publish Thread' button to
ensure that the user is logged in before attempting to publish. If the
user is not logged in, the login process is initiated and, once done,
the application sends the request to publish the thread.

Fixed a small bug where the thread is not filtered from empty array
elements after being split at the user defined points. Those empty
array elements resulted in an empty tweet component appearing in
the ThreadViewer.