Started implementing the URL splitter to protect URLs in the thread from being treated as sentences and split at full stops.
Currently, the URLs are split into their own tweets, but they need
to be combined with other tweets. I also need to figure what is
going to happen if the URL itself is way too long to fit in a tweet
and what if it is too long to be combined with other tweets.