Swapped the order of the two steps of the function that breaks text at full sentences. It now starts with splitting at newlines then resort to splitting at fullstops only if there are sentences that are too long.
The reason for the change is that splitting at newlines seems to
produce better results. Although both would indicate a full sentence,
a full stop might come in the middle of a quote, which means the
quote could be split into 2 different tweets. If the quote is too
long, then there would be no way around this. But if it's a small
quote, it would be better if it stays intact.

Also, did some cleanup work where I removed some unnecessary functions
and tests.