Added feature to allow the thread viewer to scroll when its tweets overflow.
Simplified the entities highlighting module and updated it to
support highlighting URLs. Additionally, created a React component
for the hashtag and mentions and another one for the URL to apply
custom styling to them.

Fixed a bug where the last sentence of a tweet is sometimes not
rendered when it is handled by the fullSentenceSplitter module.